add login functionality and user profile endpoint

+++ src/main/java/com/ecommerce/api/utils/SecurityUtils.java
@@ -8,7 +8,7 @@ public class SecurityUtils {
     /**
      * Ambil userId dari SecurityContext
      */
-    public static Long getCurrentUserId() {
+    public static Integer getCurrentUserId() {
         AuthenticatedUser authUser = getCurrentUser();
         return authUser != null ? authUser.getUserId() : null;
     }
